app/requirements.txt
Django
djangorestframework
celery
django-filter
django-rest-auth
requests
python-baseconv
django-phonenumber-field
phonenumbers
Pillow
django-allauth
django-extensions
gunicorn
psycopg2-binary